+++ src/main/java/leets/leenk/domain/feed/application/dto/request/FeedUploadRequest.java
@@ -4,7 +4,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
-import leets.leenk.domain.media.domain.application.dto.request.FeedMediaRequest;
+import leets.leenk.domain.media.application.dto.request.FeedMediaRequest;
 
 import java.util.List;
