+++ src/main/java/leets/leenk/domain/media/domain/service/MediaGetService.java
@@ -1,7 +1,7 @@
 package leets.leenk.domain.media.domain.service;
 
 import leets.leenk.domain.feed.domain.entity.Feed;
-import leets.leenk.domain.media.domain.application.exception.MediaNotFoundException;
+import leets.leenk.domain.media.application.exception.MediaNotFoundException;
 import leets.leenk.domain.media.domain.entity.Media;
 import leets.leenk.domain.media.domain.repository.MediaRepository;
 import lombok.RequiredArgsConstructor;
